@@ -20,61 +20,62 @@
   "bugs": "https://github.com/marmelab/admin-on-rest/issues",
   "license": "MIT",
   "devDependencies": {
-    "babel-cli": "~6.18.0",
-    "babel-core": "~6.18.2",
-    "babel-eslint": "~7.1.1",
-    "babel-loader": "~6.2.5",
+    "babel-cli": "~6.24.1",
+    "babel-core": "~6.24.1",
+    "babel-eslint": "~7.2.1",
+    "babel-loader": "~6.4.1",
     "babel-plugin-add-module-exports": "~0.2.1",
     "babel-plugin-transform-builtin-extend": "~1.1.2",
-    "babel-plugin-transform-react-jsx": "~6.8.0",
-    "babel-plugin-transform-runtime": "~6.15.0",
-    "babel-preset-es2015": "~6.18.0",
-    "babel-preset-react": "~6.16.0",
-    "babel-preset-stage-0": "~6.16.0",
-    "babel-register": "~6.18.0",
+    "babel-plugin-transform-react-jsx": "~6.24.1",
+    "babel-plugin-transform-runtime": "~6.23.0",
+    "babel-preset-es2015": "~6.24.1",
+    "babel-preset-react": "~6.24.1",
+    "babel-preset-stage-0": "~6.24.1",
+    "babel-register": "~6.24.1",
     "chromedriver": "~2.26.1",
-    "css-loader": "~0.26.0",
-    "enzyme": "~2.6.0",
-    "eslint": "~3.11.0",
-    "eslint-config-airbnb": "~13.0.0",
+    "css-loader": "~0.28.0",
+    "enzyme": "~2.8.0",
+    "eslint": "~3.19.0",
+    "eslint-config-airbnb": "~14.1.0",
     "eslint-plugin-import": "~2.2.0",
-    "eslint-plugin-jsx-a11y": "~2.2.3",
-    "eslint-plugin-react": "~6.7.1",
-    "express": "~4.14.0",
-    "extract-text-webpack-plugin": "~1.0.1",
+    "eslint-plugin-jsx-a11y": "~4.0.0",
+    "eslint-plugin-react": "~6.10.3",
+    "express": "~4.15.2",
+    "extract-text-webpack-plugin": "~2.1.0",
     "file-api": "~0.10.4",
     "full-icu": "~1.0.3",
     "ignore-styles": "~5.0.1",
     "mocha": "~3.2.0",
-    "react-addons-test-utils": "~15.4.0",
-    "selenium-webdriver": "~3.0.0-beta-3",
-    "sinon": "~1.17.6",
-    "style-loader": "~0.13.1",
+    "react-addons-test-utils": "~15.5.1",
+    "selenium-webdriver": "~3.3.0",
+    "sinon": "~2.1.0",
+    "style-loader": "~0.16.0",
     "webpack": "~1.13.2",
     "webpack-dev-server": "~1.16.2"
   },
   "dependencies": {
     "babel-runtime": "~6.18.0",
-    "inflection": "~1.10.0",
+    "inflection": "~1.12.0",
     "lodash.debounce": "~4.0.8",
     "lodash.defaultsdeep": "~4.6.0",
     "lodash.get": "~4.4.2",
     "lodash.set": "~4.3.2",
-    "material-ui": "~0.17.1",
+    "material-ui": "~0.17.3",
     "node-polyglot": "2.2.2",
+    "prop-types": "~15.5.7",
     "query-string": "~4.3.2",
-    "react": "~15.4.0",
-    "react-dom": "~15.4.0",
-    "react-dropzone": "~3.6.0",
-    "react-redux": "~4.4.5",
-    "react-router": "~4.0.0",
-    "react-router-dom": "~4.0.0",
-    "react-router-redux": "~5.0.0-alpha.4",
+    "react": "~15.5.4",
+    "react-dom": "~15.5.4",
+    "react-dropzone": "~3.12.3",
+    "react-redux": "~5.0.4",
+    "react-router": "~4.1.0",
+    "react-router-dom": "~4.1.0",
+    "react-router-redux": "~5.0.0-alpha.5",
     "react-tap-event-plugin": "~2.0.1",
-    "recompose": "~0.21.2",
+    "recompose": "~0.23.1",
     "redux": "~3.6.0",
-    "redux-form": "~6.6.1",
-    "redux-saga": "~0.14.2",
-    "reselect": "~2.5.4"
+    "redux-form": "~6.6.2",
+    "redux-saga": "~0.14.6",
+    "reselect": "~3.0.0"
   }
 }
