@@ -19,21 +19,11 @@ const CrudRoute = ({ resource, list, create, edit, show, remove, options }) => {
         </Restricted>;
     return (
         <Switch>
-            {list
-                ? <Route exact path={`/${resource}`} render={RestrictedPage(list, 'list')} />
-                : <Route path="dummy" />}
-            {create
-                ? <Route exact path={`/${resource}/create`} render={RestrictedPage(create, 'create')} />
-                : <Route path="dummy" />}
-            {edit
-                ? <Route exact path={`/${resource}/:id`} render={RestrictedPage(edit, 'edit')} />
-                : <Route path="dummy" />}
-            {show
-                ? <Route exact path={`/${resource}/:id/show`} render={RestrictedPage(show, 'show')} />
-                : <Route path="dummy" />}
-            {remove
-                ? <Route exact path={`/${resource}/:id/delete`} render={RestrictedPage(remove, 'delete')} />
-                : <Route path="dummy" />}
+            {list && <Route exact path={`/${resource}`} render={RestrictedPage(list, 'list')} />}
+            {create &&  <Route exact path={`/${resource}/create`} render={RestrictedPage(create, 'create')} />}
+            {edit && <Route exact path={`/${resource}/:id`} render={RestrictedPage(edit, 'edit')} />}
+            {show && <Route exact path={`/${resource}/:id/show`} render={RestrictedPage(show, 'show')} />}
+            {remove && <Route exact path={`/${resource}/:id/delete`} render={RestrictedPage(remove, 'delete')} />}
         </Switch>
     );
 };
