@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import MenuItem from 'material-ui/MenuItem';
 import DashboardIcon from 'material-ui/svg-icons/action/dashboard';
 import { Link } from 'react-router-dom';
