@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import inflection from 'inflection';
 import MenuItem from 'material-ui/MenuItem';
 import { Link } from 'react-router-dom';
