@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import pure from 'recompose/pure';
 import FlatButton from 'material-ui/FlatButton';
 import IconButton from 'material-ui/IconButton';
